Refactored calendar and timetable scrapers to be injected
