--- app/src/main/java/com/ice/ktuice/al/services/scraperService/exceptions/ParsingException.kt
@@ -1,3 +1,0 @@
-package com.ice.ktuice.al.services.scraperService.exceptions
-
-class ParsingException(message: String): Exception(message)
