+++ app/src/main/java/com/ice/ktuice/al/services/scrapers/base/exceptions/ParsingException.kt
@@ -1,0 +1,3 @@
+package com.ice.ktuice.al.services.scrapers.base.exceptions
+
+class ParsingException(message: String): Exception(message)
