+++ app/src/main/java/com/ice/ktuice/al/services/scrapers/calendar/CalendarHandler.kt
@@ -1,4 +1,4 @@
-package com.ice.ktuice.al.services.scraperService.ktuScraperService.handlers
+package com.ice.ktuice.al.services.scrapers.calendar
 
 import biweekly.Biweekly
 import biweekly.ICalendar
