+++ app/src/main/java/com/ice/ktuice/al/services/scrapers/calendar/CalendarScraperHanderImpl.kt
@@ -1,0 +1,10 @@
+package com.ice.ktuice.al.services.scrapers.calendar
+
+import com.ice.ktuice.models.LoginModel
+import com.ice.ktuice.models.lectureCalendarModels.CalendarModel
+
+class CalendarScraperHanderImpl: CalendarScraper {
+    override fun getCalendar(login: LoginModel): CalendarModel {
+        return CalendarHandler.getCalendar(login)
+    }
+}
